Major color rework (see description)
INstead of grays and base colors, only using base.

Chagned base to match tailwind such that lower numbers are lighter not darker

Added 700,800,900 as 300,200,100 respectively and then added new 300,200,100 values. Replaced base-100 with base-900, base-200 with base-800, and base-300 with base-700. Replaced gray- with base-

Note that base 100,200,300 are dark in the daisyui themes because that's what the website is built off of.