chore: add missing peer dependency
Svelte-Kit expects Vite to be added as a devDependency in later versions. 

In the svelte-kit version we use now it just installs the latest version stable version when running `npm install`, but in it still kinda expects you to install the latest development version as a devDependency of vite manually, as that is what svelte-kit is developed against.

So TL:DR; svelte-kit works better when you explicitly install the latest version of Vite. We need it later too.