refactor: move Dnd nav component to routes
New svelte-kit router works in such a way we can now just (by convention) have the component inside the same folder without it generating a new route. This allows us to get rid of the majority of the `lib` folder. Not that we have to, but singular stray components like this can now just be placed directly inside the routes folder.