chore!: exclude files from three view in vscode
NOTE: It literally hides the folders from the file tree in vscode. for example `node_modules` is still there, but it is not shown in the list. This because you should never touch that folder. `pnpm` should take care of it. Same goes for the `.svelte-kit`, `.build` and other should-be-hidden folders. At any time you can comment out these settings or remove them locally to get these folders to display again. This does not effect git in any way.