fix: don't split. do edge
I read the docs wrong. But I do want to test out how well edge functions work. Not sure if this costs us more money, but I assume it doesn't as it requires the usage of less ram and execution time... so if we can run edge functions, it should be cheaper, right?